pyproject.toml
[tool.poetry]
name = "gpx-record-analysis"
version = "0.1.0"
description = "Analysis of GPX records, similar to certain web services."
authors = ["Martin Ueding <mu@martin-ueding.de>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
gpxpy = "^1.4.2"
python-dateutil = "^2.8.1"
streamlit = "^0.81.0"
altair = "^4.1.0"
numpy = "^1.20.2"
matplotlib = "^3.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
